// e2e/tests/admin/cases/case-details-management/tabs/case-details-management-zgw/case-details-management-zgw-general/case-type-link.types.ts
/*
 * Copyright 2015-2026 Ritense BV, the Netherlands.
 *
 * Licensed under EUPL, Version 1.2 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" basis,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface ZaakType {
  url: string;
  omschrijving: string;
}

export interface ZaakTypeLink {
  id?: string;
  caseDefinitionKey?: string;
  caseVersionTag?: string;
  zaakTypeUrl: string;
  zakenApiPluginConfigurationId?: string;
  rsin?: string;
  createWithDossier?: boolean;
}

export interface PluginConfiguration {
  id: string;
  title: string;
}